Refactoring
Refactoring to use classes a bit more. Using modules here worked fine,
but any bigger project might want classes for the services it interacts
with. This would allow for services to set up their dependencies,
connect to dbs, without being in the main body which can cause issues on
initialiation.